refactor: test config 패키지 변경 및 테스트 검증 수정
