+++ src/test/java/com/example/jpamaster/flight/domain/repository/AirplaneRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.example.jpamaster.flight.domain.repository;
 
-import com.example.jpamaster.flight.common.annotations.ConfiguredDataJpaTest;
+import com.example.jpamaster.common.annotations.ConfiguredDataJpaTest;
 import com.example.jpamaster.flight.domain.entity.Airline;
 import com.example.jpamaster.flight.domain.entity.Airplane;
 import com.example.jpamaster.flight.domain.entity.AirplaneSeatType;
